NetworkBuffer now has valid parameter Removed NetworkBuffer variable in NetworkMessage TcpClient DataAvailable parameter type is now int32 Fixed NetworkMessage::EncododeMessage to put the size in the Buffer's header and to validate and fixed DecodeMessage to check the buffer's valid variable Fixed some methods in TcpClientWindows

+++ include/TcpClient.hpp
@@ -35,7 +35,7 @@ public:
 
 	bool Connect();
 
-	bool DataAvailable(uint16 &size);
+	bool DataAvailable(int32 &size);
 
 	//this method will receive the messages automaticaly and use the callback methods
 	void ReceiveMessages();
